+++ src/main/php/util/cmd/term/Output.class.php
@@ -1,6 +1,7 @@
 <?php namespace util\cmd\term;
 
 class Output extends \io\streams\ConsoleOutputStream {
+  public static $direct;
 
   /**
    * Write a string